fix(workflow): Show old data for project stats graph before Sep. 11th
